Add proper Tsodyks-Markram STP mechanism implementation.Not ready yet.
As can be seen from Ex7 (especially evident in the second pulse), the
state discontinuity A = A + plasticityFactor * factor seems to use the
"plasticityFactor" value from the previous timestep. This has nothing
to do with having plasticityFactor entirely calculated by a child
element, as one can get the same effect by appropriately changing the
expressions for A, B and g in the existing stpSynapse componentType.